requirements.txt
tqdm==4.67.1
Pillow==12.0.0
ttkthemes==3.3.0
PyQt5>=5.15.0
